@@ -1,36 +1,38 @@
 {
   "name": "functions",
-  "description": "Cloud Functions for Firebase",
   "scripts": {
-    "lint": "eslint .",
-    "serve": "firebase serve --only functions",
-    "shell": "firebase functions:shell",
+    "lint": "tslint --project tsconfig.json",
+    "build": "tsc",
+    "serve": "npm run build && firebase emulators:start --only functions",
+    "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log",
-    "setup-firebase": "node setupProject.js"
+    "setup-firebase": "node setupProject.js",
+    "test": "mocha -r ts-node/register --recursive 'test/**/*.test.ts' --timeout 10000 --exit"
   },
   "engines": {
-    "node": "8"
+    "node": "10"
   },
+  "main": "lib/index.js",
   "dependencies": {
-    "@google-cloud/storage": "^4.3.0",
-    "busboy": "^0.3.1",
-    "camelcase": "^5.3.1",
-    "cookie-parser": "^1.4.4",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "firebase-admin": "^8.6.0",
+    "camelcase": "^6.0.0",
+    "firebase-admin": "^9.0.0",
     "firebase-function-tools": "^1.1.3",
-    "firebase-functions": "^3.3.0",
-    "fs-extra": "^8.1.0",
-    "glob": "^7.1.6",
-    "nodemailer": "^6.4.2",
-    "sharp": "^0.25.1",
-    "uuid": "^7.0.1"
+    "firebase-functions": "^3.9.0",
+    "glob": "^7.1.6"
   },
   "devDependencies": {
-    "firebase-functions-test": "^0.1.6"
+    "@types/chai": "^4.2.12",
+    "@types/chai-as-promised": "^7.1.3",
+    "@types/mocha": "^8.0.1",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "firebase-functions-test": "^0.2.1",
+    "mocha": "^8.1.0",
+    "ts-node": "^8.10.2",
+    "tslint": "^6.1.3",
+    "typescript": "^3.9.7"
   },
   "private": true
 }
