@@ -1,9 +1,26 @@
 import React from 'react';
-import DatePicker from 'react-datepicker';
+import DatePicker, { registerLocale } from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import PropTypes from 'prop-types';
+import es from 'date-fns/locale/es';
+import en from 'date-fns/locale/en-US';
+import { shallowEqual, useSelector } from 'react-redux';
 
 import './DatePicker.scss';
+
+registerLocale('en', en);
+registerLocale('es', es);
+
+const dateFormat = locale => {
+  switch (locale) {
+    case 'en':
+      return 'MM-dd-yy';
+    case 'es':
+      return 'dd/MM/yy';
+    default:
+      return 'MM-dd-yy';
+  }
+};
 
 const DatePickerStyled = ({ name, date, setState }) => {
   const onDateChangedHandler = value =>
@@ -11,8 +28,22 @@ const DatePickerStyled = ({ name, date, setState }) => {
       ...prevState,
       [name]: value.toDateString()
     }));
-  
-  return <DatePicker selected={date} onChange={onDateChangedHandler} />;
+
+  const { locale } = useSelector(
+    state => ({
+      locale: state.preferences.locale
+    }),
+    shallowEqual
+  );
+
+  return (
+    <DatePicker
+      locale={locale}
+      dateFormat={dateFormat(locale)}
+      selected={date}
+      onChange={onDateChangedHandler}
+    />
+  );
 };
 
 DatePickerStyled.propTypes = {
