@@ -1,5 +1,4 @@
 import React from 'react';
-import { shallow } from 'enzyme';
 
 import Layout from '.';
 import NavBar from '../Navigation/NavBar';
@@ -8,13 +7,13 @@ import Footer from '../Navigation/Footer';
 
 describe('<Layout /> rendering', () => {
   it('should render without crashing', () => {
-    const component = shallow(<Layout>Test</Layout>);
+    const { component } = shallowWithProviders(<Layout>Test</Layout>)({});
 
     expect(component).toMatchSnapshot();
   });
 
   it('should render <NavBar /> component correctly', () => {
-    const { component } = mountWithProvider(<Layout>Test</Layout>)({
+    const { component } = mountWithProviders(<Layout>Test</Layout>)({
       auth: {
         userData: {}
       }
@@ -24,7 +23,7 @@ describe('<Layout /> rendering', () => {
   });
 
   it('should render <Aside /> component correctly', () => {
-    const { component } = mountWithProvider(<Layout>Test</Layout>)({
+    const { component } = mountWithProviders(<Layout>Test</Layout>)({
       auth: {
         userData: {}
       }
@@ -34,7 +33,7 @@ describe('<Layout /> rendering', () => {
   });
 
   it('should render <Footer /> component correctly', () => {
-    const { component } = mountWithProvider(<Layout>Test</Layout>)({
+    const { component } = mountWithProviders(<Layout>Test</Layout>)({
       auth: {
         userData: {}
       }
@@ -44,7 +43,7 @@ describe('<Layout /> rendering', () => {
   });
 
   it('should render a div with the children', () => {
-    const { component } = mountWithProvider(<Layout>Test</Layout>)({
+    const { component } = mountWithProviders(<Layout>Test</Layout>)({
       auth: {
         userData: {}
       }
