@@ -1,12 +1,14 @@
 import React from 'react';
 import { Redirect } from 'react-router-dom';
+import * as reactRedux from 'react-redux';
 
+import * as authActions from 'state/actions/auth';
 import paths from '../Router/paths';
 import Login from '.';
 
 describe('<Login /> rendering', () => {
   it('should render without crashing', () => {
-    const { component } = mountWithProvider(<Login />)({
+    const { component } = mountWithProviders(<Login />)({
       auth: {
         userData: {}
       }
@@ -16,7 +18,7 @@ describe('<Login /> rendering', () => {
   });
 
   it('should display an error message when there is an error', () => {
-    const { component } = mountWithProvider(<Login />)({
+    const { component } = mountWithProviders(<Login />)({
       auth: {
         userData: {},
         error: 'sample error'
@@ -27,7 +29,7 @@ describe('<Login /> rendering', () => {
   });
 
   it('should redirect to /home when the user is authenticated', () => {
-    const { component } = mountWithProvider(<Login />)({
+    const { component } = mountWithProviders(<Login />)({
       auth: {
         userData: {
           id: 'some userId'
@@ -37,4 +39,49 @@ describe('<Login /> rendering', () => {
 
     expect(component.contains(<Redirect to={paths.ROOT} />)).toEqual(true);
   });
+});
+
+const dispatchMock = jest.fn();
+
+beforeEach(() => {
+  jest.spyOn(reactRedux, 'useDispatch').mockImplementation(() => dispatchMock);
+  jest.spyOn(authActions, 'authFacebook').mockImplementation(jest.fn);
+  jest.spyOn(authActions, 'authGoogle').mockImplementation(jest.fn);
+  jest.spyOn(authActions, 'authMicrosoft').mockImplementation(jest.fn);
+});
+
+it('should dispatch authFacebook action when the user tries to log in with Facebook', () => {
+  const { component } = mountWithProviders(<Login />)({
+    auth: {
+      userData: {}
+    }
+  });
+
+  component.find('.is-facebook').simulate('click');
+
+  expect(authActions.authFacebook).toHaveBeenCalled();
+});
+
+it('should dispatch authGoogle action when the user tries to log in with Google', () => {
+  const { component } = mountWithProviders(<Login />)({
+    auth: {
+      userData: {}
+    }
+  });
+
+  component.find('.is-google').simulate('click');
+
+  expect(authActions.authGoogle).toHaveBeenCalled();
+});
+
+it('should dispatch authMicrosoft action when the user tries to log in with Microsoft', () => {
+  const { component } = mountWithProviders(<Login />)({
+    auth: {
+      userData: {}
+    }
+  });
+
+  component.find('.is-microsoft').simulate('click');
+
+  expect(authActions.authMicrosoft).toHaveBeenCalled();
 });
