@@ -1,7 +1,7 @@
 import React from 'react';
 import * as reactRedux from 'react-redux';
 
-import * as actions from '../../state/actions/auth';
+import * as actions from 'state/actions/auth';
 import Router from '.';
 
 describe('<Router /> rendering', () => {
@@ -15,7 +15,7 @@ describe('<Router /> rendering', () => {
   });
 
   it('should render without crashing', () => {
-    const { component } = shallowWithProvider(<Router />)({
+    const { component } = shallowWithProviders(<Router />)({
       auth: {}
     });
 
