@@ -1,13 +1,17 @@
 import React from 'react';
 
+import { useFormatMessage } from 'hooks';
+
 const Submenu = () => (
   <>
     <section className="hero is-hero-bar">
       <div className="hero-body">
-        <h1 className="title">Submenu</h1>
+        <h1 className="title">{useFormatMessage('Submenu.submenu')}</h1>
       </div>
     </section>
-    <section className="section is-main-section">Submenu content</section>
+    <section className="section is-main-section">
+      {useFormatMessage('Submenu.content')}
+    </section>
   </>
 );
 
