@@ -1,10 +1,8 @@
 import { createAction } from 'redux-act';
-import uuid from 'uuid/v4';
 import { toastr } from 'react-redux-toastr';
 
-import axios from '../../utils/axios';
-import firebase from '../../firebase';
-import { firebaseError } from '../../utils';
+import { firebaseError } from 'utils';
+import firebase from 'firebase.js';
 import { checkUserData, AUTH_UPDATE_USER_DATA } from './auth';
 
 export const USERS_FETCH_DATA_INIT = createAction('USERS_FETCH_DATA_INIT');
@@ -35,23 +33,23 @@ export const USERS_MODIFY_USER_FAIL = createAction('USERS_MODIFY_USER_FAIL');
 
 export const USERS_CLEAN_UP = createAction('USERS_CLEAN_UP');
 
+export const USERS_CLEAR_DATA_LOGOUT = createAction('USERS_CLEAR_DATA_LOGOUT');
+
 export const fetchUsers = () => {
   return async (dispatch, getState) => {
     dispatch(checkUserData());
 
     dispatch(USERS_FETCH_DATA_INIT());
 
-    const { tenant, id } = getState().auth.userData;
+    const { id } = getState().auth.userData;
 
     let users;
 
     try {
-      const ref = firebase.database().ref('users');
-
       users = (
-        await ref
-          .orderByChild('tenant')
-          .equalTo(tenant)
+        await firebase
+          .database()
+          .ref('users')
           .once('value')
       ).val();
     } catch (error) {
@@ -74,19 +72,46 @@ export const fetchUsers = () => {
   };
 };
 
+const deleteLogo = oldLogo => {
+  if (!oldLogo.includes('firebasestorage')) {
+    return null;
+  }
+  const logoPath = oldLogo
+    .split('users%2F')
+    .pop()
+    .split('?alt=media')
+    .shift();
+  return firebase
+    .storage()
+    .ref(`users/${logoPath}`)
+    .delete();
+};
+
 export const deleteUser = id => {
-  return async dispatch => {
+  return async (dispatch, getState) => {
     dispatch(USERS_DELETE_USER_INIT());
+    const { locale } = getState().preferences;
+    const { logoUrl } = getState()
+      .users.data.filter(user => user.id === id)
+      .pop();
 
-    const user = firebase.auth().currentUser;
+    const deleteLogoTask = logoUrl ? deleteLogo(logoUrl) : null;
 
-    const userToken = await user.getIdToken();
+    const deleteUserTask = firebase
+      .database()
+      .ref(`users/${id}`)
+      .remove();
 
     try {
-      await axios(userToken).delete(`/users/${id}`);
+      await Promise.all([deleteLogoTask, deleteUserTask]);
     } catch (error) {
-      toastr.error('', error);
-      return dispatch(USERS_DELETE_USER_FAIL({ error }));
+      const errorMessage = firebaseError(error.code, locale);
+      toastr.error('', errorMessage);
+      return dispatch(
+        USERS_DELETE_USER_FAIL({
+          error: errorMessage
+        })
+      );
     }
 
     toastr.success('', 'The user was deleted.');
@@ -100,6 +125,30 @@ export const clearUsersData = () => {
   };
 };
 
+export const clearUsersDataLogout = () => {
+  return dispatch => {
+    dispatch(USERS_CLEAR_DATA_LOGOUT());
+  };
+};
+
+const uploadLogo = (uid, file) => {
+  const storageRef = firebase.storage().ref();
+
+  const fileExtension = file.name.split('.').pop();
+
+  const fileName = `${uid}.${fileExtension}`;
+
+  return storageRef.child(`users/${fileName}`).put(file);
+};
+
+const getLogoUrl = (uid, file) => {
+  const fileExtension = file.name.split('.').pop();
+
+  const bucketUrl = `${process.env.REACT_APP_FIRE_BASE_STORAGE_API}`;
+
+  return `${bucketUrl}/o/users%2F${uid}_200x200.${fileExtension}?alt=media`;
+};
+
 export const createUser = ({
   name,
   email,
@@ -110,71 +159,66 @@ export const createUser = ({
 }) => {
   return async (dispatch, getState) => {
     dispatch(USERS_CREATE_USER_INIT());
+    const { locale } = getState().preferences;
 
-    const { tenant } = getState().auth.userData;
+    let response;
+    try {
+      const createUserAuth = firebase
+        .functions()
+        .httpsCallable('httpsCreateUser');
 
-    const user = firebase.auth().currentUser;
-
-    const userToken = await user.getIdToken();
-
-    const body = new FormData();
-
-    if (file) {
-      const fileExtension = file.name.split('.')[1];
-
-      const fileName = `${uuid()}.${fileExtension}`;
-
-      body.append('logo', file, fileName);
+      response = await createUserAuth({ email, isAdmin });
+    } catch (error) {
+      const errorMessage = firebaseError(error.message, locale);
+      toastr.error('', errorMessage);
+      return dispatch(
+        USERS_CREATE_USER_FAIL({
+          error: errorMessage
+        })
+      );
     }
 
-    body.append('name', name);
-    body.append('location', location);
-    body.append('email', email);
-    body.append('password', uuid());
-    body.append('tenant', tenant);
-    body.append('createdAt', createdAt);
-    body.append('isAdmin', isAdmin);
+    const { uid } = response.data;
 
-    axios(userToken)
-      .post('/users', body)
-      .then(async response => {
-        const userCreated = response.data;
+    let uploadLogoTask = null;
+    let logoUrl = null;
+    if (file) {
+      logoUrl = getLogoUrl(uid, file);
+      uploadLogoTask = uploadLogo(uid, file);
+    }
 
-        const actionCodeSettings = {
-          url: process.env.REACT_APP_LOGIN_PAGE_URL,
-          handleCodeInApp: true
-        };
+    const createUserDbTask = firebase
+      .database()
+      .ref(`users/${uid}`)
+      .set({ name, email, location, logoUrl, createdAt, isAdmin });
 
-        try {
-          await firebase
-            .auth()
-            .sendSignInLinkToEmail(email, actionCodeSettings);
-        } catch (error) {
-          const errorMessage = firebaseError(error.response.data.error.code);
+    const actionCodeSettings = {
+      url: process.env.REACT_APP_LOGIN_PAGE_URL,
+      handleCodeInApp: true
+    };
 
-          return dispatch(
-            USERS_CREATE_USER_FAIL({
-              error: errorMessage
-            })
-          );
-        }
+    const sendSignInLinkToEmailTask = firebase
+      .auth()
+      .sendSignInLinkToEmail(email, actionCodeSettings);
 
-        toastr.success('', 'User created successfully');
-        return dispatch(
-          USERS_CREATE_USER_SUCCESS({
-            user: userCreated
-          })
-        );
-      })
-      .catch(error => {
-        const errorMessage = firebaseError(error.response.data.error.code);
-        toastr.error('', errorMessage);
-        return dispatch(
-          USERS_CREATE_USER_FAIL({
-            error: errorMessage
-          })
-        );
-      });
+    try {
+      await Promise.all([
+        uploadLogoTask,
+        createUserDbTask,
+        sendSignInLinkToEmailTask
+      ]);
+    } catch (error) {
+      const errorMessage = firebaseError(error.code, locale);
+      toastr.error('', errorMessage);
+      return dispatch(
+        USERS_CREATE_USER_FAIL({
+          error: errorMessage
+        })
+      );
+    }
+
+    toastr.success('', 'User created successfully');
+    return dispatch(USERS_CREATE_USER_SUCCESS({ user: response.data }));
   };
 };
 
@@ -190,60 +234,60 @@ export const modifyUser = ({
 }) => {
   return async (dispatch, getState) => {
     dispatch(USERS_MODIFY_USER_INIT());
+    const { locale } = getState().preferences;
+    const { logoUrl } = isProfile
+      ? getState().auth.userData
+      : getState()
+          .users.data.filter(user => user.id === id)
+          .pop();
 
-    const { tenant } = getState().auth.userData;
-
-    const user = firebase.auth().currentUser;
-
-    const userToken = await user.getIdToken();
-
-    const body = new FormData();
-
+    let deleteLogoTask;
+    let uploadLogoTask;
+    let newLogoUrl = null;
     if (file) {
-      const fileExtension = file.name.split('.')[1];
-
-      const fileName = `${uuid()}.${fileExtension}`;
-
-      body.append('logo', file, fileName);
+      newLogoUrl = getLogoUrl(id, file);
+      deleteLogoTask = logoUrl && deleteLogo(logoUrl);
+      uploadLogoTask = uploadLogo(id, file);
     }
 
-    body.append('name', name);
-    body.append('location', location);
-    body.append('tenant', tenant);
-    body.append('createdAt', createdAt);
-    body.append('isAdmin', isAdmin);
+    const userData = {
+      name,
+      location,
+      createdAt,
+      isAdmin,
+      logoUrl: logoUrl || newLogoUrl
+    };
 
-    axios(userToken)
-      .patch(`/users/${id}`, body)
-      .then(response => {
-        const userCreated = response.data;
-        const { uid } = firebase.auth().currentUser;
+    const updateUserDbTask = firebase
+      .database()
+      .ref(`users/${id}`)
+      .update(userData);
 
-        if (id === uid) {
-          dispatch(AUTH_UPDATE_USER_DATA({ ...userCreated }));
-        }
+    const { uid } = firebase.auth().currentUser;
 
-        if (isProfile) {
-          toastr.success('', 'Profile updated successfully');
-        } else if (isEditing) {
-          toastr.success('', 'User updated successfully');
-        }
+    if (id === uid) {
+      dispatch(AUTH_UPDATE_USER_DATA({ ...userData, id }));
+    }
 
-        return dispatch(
-          USERS_MODIFY_USER_SUCCESS({
-            user: userCreated
-          })
-        );
-      })
-      .catch(error => {
-        const errorMessage = firebaseError(error.response.data.error.code);
-        toastr.error('', errorMessage);
-        return dispatch(
-          USERS_MODIFY_USER_FAIL({
-            error: errorMessage
-          })
-        );
-      });
+    try {
+      await Promise.all([deleteLogoTask, uploadLogoTask, updateUserDbTask]);
+    } catch (error) {
+      const errorMessage = firebaseError(error.code, locale);
+      toastr.error('', errorMessage);
+      return dispatch(
+        USERS_MODIFY_USER_FAIL({
+          error: errorMessage
+        })
+      );
+    }
+
+    if (isProfile) {
+      toastr.success('', 'Profile updated successfully');
+    } else if (isEditing) {
+      toastr.success('', 'User updated successfully');
+    }
+
+    return dispatch(USERS_MODIFY_USER_SUCCESS({ user: { ...userData, id } }));
   };
 };
 
