@@ -1,3 +1,10 @@
+import { createIntl, createIntlCache } from 'react-intl';
+
+import english from 'languages/en';
+import spanish from 'languages/es';
+import en from 'assets/en.png';
+import es from 'assets/es.png';
+
 export const FIREBASE_RESPONSE = {
   EMAIL_IN_USE: 'auth/email-already-exists',
   EMAIL_INVALID: 'auth/invalid-email',
@@ -6,44 +13,34 @@ export const FIREBASE_RESPONSE = {
   USER_DISABLED: 'auth/user-disabled',
   TOO_MANY_REQUESTS: 'auth/too-many-requests',
   EXPIRED_ACTION_CODE: 'auth/expired-action-code',
-  INVALID_ACTION_CODE: 'auth/invalid-action-code'
+  INVALID_ACTION_CODE: 'auth/invalid-action-code',
+  QUOTA_EXCEEDED_STORAGE: 'storage/quota-exceeded',
+  UNAUTHENTICATED_STORAGE: 'storage/unauthenticated',
+  UNAUTHORIZED_STORAGE: 'storage/unauthorized'
 };
 
-export const firebaseError = error => {
-  let errorMessage = '';
+export const messages = {
+  en: english,
+  es: spanish
+};
 
-  switch (error) {
-    case FIREBASE_RESPONSE.EMAIL_IN_USE:
-      errorMessage = 'Email already in use';
-      break;
-    case FIREBASE_RESPONSE.EMAIL_INVALID:
-      errorMessage = 'Email is invalid';
-      break;
-    case FIREBASE_RESPONSE.EMAIL_NOT_FOUND:
-      errorMessage = 'Invalid credentials';
-      break;
-    case FIREBASE_RESPONSE.PASSWORD_INVALID:
-      errorMessage = 'Invalid credentials';
-      break;
-    case FIREBASE_RESPONSE.USER_DISABLED:
-      errorMessage = 'User disabled';
-      break;
-    case FIREBASE_RESPONSE.TOO_MANY_REQUESTS:
-      errorMessage = 'Too many attempts made, try again later';
-      break;
-    case FIREBASE_RESPONSE.EXPIRED_ACTION_CODE:
-      errorMessage =
-        'The invitation link has expired, get in touch with your administrator';
-      break;
-    case FIREBASE_RESPONSE.INVALID_ACTION_CODE:
-      errorMessage =
-        'The invitation link has expired, get in touch with your administrator';
-      break;
-    default:
-      errorMessage = 'Unknown error, get in touch with your administrator';
-  }
+const getIntlContext = locale => {
+  const cache = createIntlCache();
+  return createIntl(
+    {
+      locale,
+      messages: messages[locale]
+    },
+    cache
+  );
+};
 
-  return errorMessage;
+export const firebaseError = (error, locale) => {
+  const intl = getIntlContext(locale);
+  return intl.formatMessage({
+    id: error,
+    defaultMessage: messages[locale]['utils.default']
+  });
 };
 
 export const validateEmail = email => {
@@ -53,7 +50,7 @@ export const validateEmail = email => {
   );
 };
 
-export const inputValidations = (email, password) => {
+export const inputValidations = (email, password, locale) => {
   let inputs = {
     email: {
       modifier: null,
@@ -65,6 +62,7 @@ export const inputValidations = (email, password) => {
     },
     canSubmit: null
   };
+  const intl = getIntlContext(locale);
 
   const setInputs = (key, value) => {
     inputs = { ...inputs, [`${key}`]: value };
@@ -75,7 +73,7 @@ export const inputValidations = (email, password) => {
   if (email && !isValidEmail) {
     setInputs('email', {
       modifier: 'is-danger',
-      message: 'Invalid email'
+      message: intl.formatMessage({ id: 'utils.invalidEmail' })
     });
   }
 
@@ -84,12 +82,12 @@ export const inputValidations = (email, password) => {
   if (isValidPassword) {
     setInputs('password', {
       modifier: 'is-success',
-      message: 'Safe password'
+      message: intl.formatMessage({ id: 'utils.safePassword' })
     });
   } else if (password) {
     setInputs('password', {
       modifier: 'is-danger',
-      message: 'Unsafe password'
+      message: intl.formatMessage({ id: 'utils.unsafePassword' })
     });
   }
 
@@ -98,4 +96,13 @@ export const inputValidations = (email, password) => {
   }
 
   return inputs;
+};
+
+export const availableLocales = Object.keys(messages);
+
+export const browserLocale = navigator.language.split(/[-_]/)[0];
+
+export const flags = {
+  en,
+  es
 };
